Check if files exist in import paths in `normalize_import_paths`
When defining multiple include paths and referencing files that exist in
a subdirectory of one of those import paths, `normalize_import_paths`
would produce paths relative to one of the include paths.

This would then not be findable later in `generate/2` which would look
for that file (now with a relative path!) in the list of files found in
`request.proto_file`, which are not relative to include paths if they
are found within those include paths.

This in turn would lead to `ProtobufGenerate.CodeGen.generate/3` being
called with a `nil` file descriptor, which fails.

The solution is to only use relative paths if the file path defined is
not extant under any of the include paths. If it is, leave it as it is,
as it will be found under the `include_path`s.